refactor AdvanceTextEditor name to AdvancedTextEditor and related changes
