+++ components/advanced_text_editor/advanced_text_editor.scss
