+++ components/advanced_text_editor/formatting_bar/formatting_bar.tsx
