+++ components/advanced_text_editor/formatting_bar/index.ts
