+++ components/advanced_text_editor/toggle_formatting_bar/toggle_formatting_bar.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 import classNames from 'classnames';
 
 import Constants from 'utils/constants';
-import KeyboardShortcutSequence, {KEYBOARD_SHORTCUTS} from '../../keyboard_shortcuts/keyboard_shortcuts_sequence';
+import KeyboardShortcutSequence, {KEYBOARD_SHORTCUTS} from 'components/keyboard_shortcuts/keyboard_shortcuts_sequence';
 import OverlayTrigger from 'components/overlay_trigger';
 import Tooltip from 'components/tooltip';
